requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
langchain-classic

faiss-cpu
pypdf
python-dotenv
pydantic